Extract and verify fbsr cookie
Another step for supporting OAuth 2.0. If the user exists in the DB, everything
now works -- but the call to get the user's name from Facebook in order to
create that DB entry fails, since the "code" we now have isn't an access token.
Need to call another endpoint to get that token.